requirements.txt
pandas
kaggle
textblob
spacy
nltk
pandas
scikit-learn
numpy
gensim
matplotlib
seaborn
tensorflow
transformers
openai
python-dotenv
flask
chromadb
langchain
tiktoken
unstructured
langchain_community
pinecone
pinecone-client
pypdf
huggingface_hub
wikipedia
numexpr
langchain_experimental
watermark
tabulate
faiss-cpu
einops
accelerate
bitsandbytes
datasets
loralib
sentencepiece
tokenizers
tf-keras
langchain-google-genai
